Big Housekeep III: filters > home, rename Explore components
