+++ src/components/context/ProvidersWrap.tsx
@@ -2,10 +2,11 @@ import React, { FC } from 'react'
 import { CssBaseline } from '@material-ui/core'
 import { ThemeProvider } from '@material-ui/styles'
 
-import './style.css'
+import '../style.css'
 
-import { GlobalProvider, SymbAndLabelProvider } from './context'
-import { theme } from './config/theme'
+import { GlobalProvider } from './GlobalContext'
+import { SymbAndLabelProvider } from './SymbAndLabelContext'
+import { theme } from '../config/theme'
 
 // Everything the app needs except Routes. This makes it testable and reusable
 // (e.g. <MemoryRouter>)
