+++ src/components/context/index.ts
@@ -1,3 +1,4 @@
 export * from './GlobalContext'
+export * from './ProvidersWrap'
 export * from './SymbAndLabelContext'
 export * from './types'
