+++ src/components/explore/index.ts
@@ -1,3 +1,7 @@
 export * from './Breadcrumbs'
 export * from './Explore'
-export * from './Sift'
+export * from './ExploreSubView'
+export * from './LangCardsList'
+export * from './MidLevelExplore'
+export * from './config'
+export * from './types'
