+++ src/components/generic/icons-and-swatches/index.ts
@@ -1,0 +1,2 @@
+export * from './FlagWithTitle'
+export * from './SwatchOrFlagOrIcon'
