+++ src/components/home/ClearFiltersBtn.tsx
